third_party/blink/web_tests/svg/dynamic-updates/SVGImageElement-svgdom-preserveAspectRatio-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'preserveAspectRatio' property of the SVGImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageElement.preserveAspectRatio.baseVal.align is SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMAXYMAX
PASS imageElement.preserveAspectRatio.baseVal.meetOrSlice is SVGPreserveAspectRatio.SVG_MEETORSLICE_MEET
PASS imageElement.preserveAspectRatio.baseVal.align is SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_NONE
PASS imageElement.preserveAspectRatio.baseVal.meetOrSlice is SVGPreserveAspectRatio.SVG_MEETORSLICE_MEET
PASS successfullyParsed is true

TEST COMPLETE
